Split check runner phases
